feat: remove `Result` from `assign_advice` return value
* apply `cargo clippy --fix`

+++ halo2_proofs/src/dev/failure.rs
@@ -105,10 +105,10 @@ impl FailureLocation {
             })
             .map(|(r_i, r)| FailureLocation::InRegion {
                 region: (r_i, r.name.clone()).into(),
-                offset: failure_row as usize - r.rows.unwrap().0 as usize,
+                offset: failure_row - r.rows.unwrap().0,
             })
             .unwrap_or_else(|| FailureLocation::OutsideRegion {
-                row: failure_row as usize,
+                row: failure_row,
             })
     }
 }
